extract: default destination folder fix
According to README.md:

To extract data from a J2A file:

python j2a-extract.py <anims_file>

the output folder will be in the same folder as the .J2A file, with the dot swapped for a dash; for instance if you call this with C:\Games\Jazz2\Anims.j2a, the output will be C:\Games\Jazz2\Anims-j2a\.

However, j2a-extract.py was creating set-000 (and any other set-### folders) in the working folder by default instead. This is unlikely to be useful, violates the description in the readme, and goes exceptionally poorly if the user tries to extract two or more different .j2a files in the same working folder.

This commit restores the proper default behavior.