extract: decode/render 32-bit sprites
Assuming that 32-bit sprites use the same RLE as 8-bit sprites, but with a 0x8000 bit set on the Height short to mark that they are 32-bit, these edits ensure they are read properly as tuples. A new render_bitdepth_appropriate_pixelmap method handles choosing whether to render any given frame as a paletted .png or a 32-bit .png depending on its actual contents, and this is used by j2a-extract.py in place of the monolithic render_paletted_pixelmap.