import: final(?) touches
Generate the mask based on pixel opacity!=0, which may be not be ideal but is better than ALL pixels being solid as it was before. Also add support for RGB and Grayscale("L") images, treated as RGBA and paletted("P") respectively.

(Grayscale sprites are common when working with BLEND::CUSTOM_ALPHA.)